Moved sorting functions into Report class
- Removed ReportSorter Interface and moved its functionality into Report class
- Implemented Report class tests
